Rename engineDBTypes.tsx to engineTypes.tsx as that is more representative of it's contents
