+++ src/renderer/state/dataset-list/model/reducerTypes.ts
@@ -1,4 +1,4 @@
-import { Dataset } from '_/renderer/view/helpers/constants/engineDBTypes';
+import { Dataset } from '_/renderer/view/helpers/constants/engineTypes';
 
 /**
  * State for the list of datasets.
