+++ src/renderer/state/delete-modals/model/reducerTypes.ts
@@ -1,4 +1,4 @@
-import { Dataset } from '_/renderer/view/helpers/constants/engineDBTypes';
+import { Dataset } from '_/renderer/view/helpers/constants/engineTypes';
 
 /**
  * State for whether the delete dataset dialog is open or not along with the dataset it should display.
