+++ src/renderer/view/components/delete-dataset/DeleteDataset.tsx
@@ -15,7 +15,7 @@ import { connect } from 'react-redux';
 import { IOpenCloseDeleteDatasetReducer } from '_/renderer/state/delete-modals/model/reducerTypes';
 import { openCloseDeleteDatasetAction } from '_/renderer/state/delete-modals/DeleteModalsActions';
 import { IAppState } from '_/renderer/state/reducers';
-import { Dataset, nullDataset } from '../../helpers/constants/engineDBTypes';
+import { Dataset, nullDataset } from '../../helpers/constants/engineTypes';
 import { IOpenCloseDeleteDatasetAction } from '_/renderer/state/delete-modals/model/actionTypes';
 import { EngineRequestHandler } from '_/renderer/engine-requests/engineRequestHandler';
 import { toast, waitTillEngineJobComplete } from '../../helpers/functionHelpers';
