+++ src/renderer/view/pages/Models.tsx
@@ -7,7 +7,7 @@ import { connect } from 'react-redux';
 
 import { EngineRequestHandler } from '_/renderer/engine-requests/engineRequestHandler';
 import { ModelCard } from '../components/model-card/ModelCard';
-import { Model, nullModel } from '../helpers/constants/engineDBTypes';
+import { Model, nullModel } from '../helpers/constants/engineTypes';
 import { useVerticalScrollbar } from '_/renderer/view/helpers/hooks/scrollbar';
 import { changeSelectedPageAction } from '_/renderer/state/side-menu/SideModelAction';
 import { IChangeSelectedPageAction } from '_/renderer/state/side-menu/model/actionTypes';
